requirements.txt
asgiref==3.4.1
astroid==2.7.2
attrs==21.2.0
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
drf-spectacular==0.18.0
inflection==0.5.1
isort==5.9.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.2.0
psycopg2==2.9.1
pylint==2.10.2
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
wrapt==1.12.1
